Discuss screen login auth and posts auth setup

+++ backend/routes/discussRoutes.js
@@ -2,6 +2,7 @@ const { json } = require("express");
 const express = require("express");
 const router = express.Router();
 
+const protect = require("../middleware/authMiddleware");
 const Post = require("../models/post");
 
 // @desc    Get all posts in descending order of created time
@@ -20,12 +21,12 @@ router.get("/", async (req, res) => {
 // @desc    Add a post
 // @route   POST /api/discuss
 // @access  Public
-router.post("/", async (req, res) => {
+router.post("/", protect, async (req, res) => {
   try {
-    const { userName, createdAt, title, description, sortingDate } = req.body;
+    const { createdAt, title, description, sortingDate } = req.body;
 
     const newPost = new Post({
-      userName,
+      user: req.user._id,
       createdAt,
       description,
       title,
@@ -42,7 +43,7 @@ router.post("/", async (req, res) => {
 // @desc    Add a reply
 // @route   POST /api/discuss/:id
 // @access  Public
-router.post("/:id", async (req, res) => {
+router.post("/:id", protect, async (req, res) => {
   try {
     const post = await Post.findById(req.params.id);
 
@@ -50,16 +51,16 @@ router.post("/:id", async (req, res) => {
       return res.status(404).json({ message: "Post not found" });
     }
 
-    const {
-      replyUserName,
-      replyCreatedAt,
-      replyDescription,
-      replySortingDate,
-    } = req.body;
+    const { replyCreatedAt, replyDescription, replySortingDate } = req.body;
 
     post.replies = [
       ...post.replies,
-      { replyUserName, replyCreatedAt, replyDescription, replySortingDate },
+      {
+        replyUser: req.user._id,
+        replyCreatedAt,
+        replyDescription,
+        replySortingDate,
+      },
     ];
 
     const result = await Post.findByIdAndUpdate(
@@ -79,7 +80,7 @@ router.post("/:id", async (req, res) => {
 // @desc    Delete a post
 // @route   DELETE /api/discuss/:id
 // @access  Public
-router.delete("/:id", async (req, res) => {
+router.delete("/:id", protect, async (req, res) => {
   try {
     const exists = await Post.findById(req.params.id);
 
@@ -87,9 +88,12 @@ router.delete("/:id", async (req, res) => {
       return res.status(404).json({ message: "Post not found" });
     }
 
-    await Post.findByIdAndRemove(req.params.id);
-
-    res.json({ message: "Deleted post successfully" });
+    if (String(exists.user._id) === String(req.user._id)) {
+      await Post.findByIdAndRemove(req.params.id);
+      res.json({ message: "Deleted post successfully" });
+    } else {
+      res.status(400).json({ message: "The post dosen't belong to this user" });
+    }
   } catch (error) {
     res.status(500).json({ message: error.message });
   }
@@ -98,7 +102,7 @@ router.delete("/:id", async (req, res) => {
 // @desc    Delete a reply
 // @route   DELETE /api/discuss/reply/:id
 // @access  Public
-router.delete("/reply/:postId/:replyId", async (req, res) => {
+router.delete("/reply/:postId/:replyId", protect, async (req, res) => {
   try {
     const post = await Post.findById(req.params.postId);
 
@@ -106,21 +110,27 @@ router.delete("/reply/:postId/:replyId", async (req, res) => {
       return res.status(404).json({ message: "Post dosen't exist" });
     }
 
-    const { replies } = post;
+    if (String(post.user._id) === String(req.user._id)) {
+      const { replies } = post;
 
-    var updatedReplies = replies.filter(
-      (reply) => String(reply._id) !== String(req.params.replyId)
-    );
+      var updatedReplies = replies.filter(
+        (reply) => String(reply._id) !== String(req.params.replyId)
+      );
 
-    const result = await Post.findByIdAndUpdate(
-      req.params.postId,
-      {
-        $set: { replies: updatedReplies },
-      },
-      { new: true }
-    );
+      const result = await Post.findByIdAndUpdate(
+        req.params.postId,
+        {
+          $set: { replies: updatedReplies },
+        },
+        { new: true }
+      );
 
-    res.json(result);
+      res.json(result);
+    } else {
+      res
+        .status(400)
+        .json({ message: "The reply dosen't belong to this user" });
+    }
   } catch (error) {
     res.status(500).json({ message: error.message });
   }
